BGE: Moto: Use float for MT_Scalar.
Test:
  4104 rotating spheres.
  Without moto modifications:
    Logic : 5.15 : 5%
    SceneGraph : 8.50 : 8%
    Rasterizer : 20.90 : 21%
  With:
    Logic : 5.00 : 4%
    SceneGraph : 8.00 : 8%
    Rasterizer : 20.25 : 20%

+++ source/gameengine/Physics/Bullet/CcdPhysicsEnvironment.cpp
@@ -1455,7 +1455,7 @@ struct OcclusionBuffer
 		m[14] = btScalar(m1[ 2]*m2[12]+m1[ 6]*m2[13]+m1[10]*m2[14]+m1[14]*m2[15]);
 		m[15] = btScalar(m1[ 3]*m2[12]+m1[ 7]*m2[13]+m1[11]*m2[14]+m1[15]*m2[15]);
 	}
-	void		setup(int size, const int *view, double modelview[16], double projection[16])
+	void		setup(int size, const int *view, float modelview[16], float projection[16])
 	{
 		m_initialized=false;
 		m_occlusion=false;
@@ -1505,7 +1505,7 @@ struct OcclusionBuffer
 		m_initialized = true;
 		m_occlusion = false;
 	}
-	void		SetModelMatrix(double *fl)
+	void		SetModelMatrix(float *fl)
 	{
 		CMmat4mul(m_mtc,m_wtc,fl);
 		if (!m_initialized)
@@ -1896,7 +1896,7 @@ struct	DbvtCullingCallback : btDbvt::ICollide
 			KX_GameObject* gameobj = KX_GameObject::GetClientObject(info);
 			if (gameobj && gameobj->GetOccluder())
 			{
-				double* fl = gameobj->GetOpenGLMatrixPtr()->getPointer();
+				float* fl = gameobj->GetOpenGLMatrixPtr()->getPointer();
 				// this will create the occlusion buffer if not already done
 				// and compute the transformation from model local space to clip space
 				m_ocb->SetModelMatrix(fl);
@@ -1937,7 +1937,7 @@ struct	DbvtCullingCallback : btDbvt::ICollide
 };
 
 static OcclusionBuffer gOcb;
-bool CcdPhysicsEnvironment::CullingTest(PHY_CullingCallback callback, void* userData, MT_Vector4 *planes, int nplanes, int occlusionRes, const int *viewport, double modelview[16], double projection[16])
+bool CcdPhysicsEnvironment::CullingTest(PHY_CullingCallback callback, void* userData, MT_Vector4 *planes, int nplanes, int occlusionRes, const int *viewport, float modelview[16], float projection[16])
 {
 	if (!m_cullingTree)
 		return false;
